לוגיקה לכניסה

+++ src/project/Database/DBInserts.java
@@ -72,7 +72,7 @@ public class DBInserts {
         return dbInitializer.connectAndExecute(sql);
     }
 
-    public String insertForPatient (String id, String pswd, String name, String lname, String email, double wieght, int height, String bdate, String gender,String phoneNumber) {
+    public String insertForPatient (String id, String pswd, String name, String lname, String email, double wieght, double height, String bdate, String gender,String phoneNumber) {
         String sql="INSERT INTO users(id, password, firstname, lastname, email, weight, height, birthdate, userRole, gender, phoneNumber)" +
                 "VALUES ('"+id+"', '"+pswd+"', '"+name+"', '"+lname+"', '"+email+"', '"+wieght+"', '"+height+"', '"+bdate+"', 'P','"+gender +"', '"+phoneNumber+"');";
         return dbInitializer.connectAndExecute(sql);
